Fixes from ATM CR
